+++ WultraMobileTokenSDK/Operations/Service/WMTOperationsImpl.swift
@@ -131,7 +131,7 @@ class WMTOperationsImpl<T: WMTUserOperation>: WMTOperations, WMTService {
     private var isPollingPaused: Bool { return pollingTimer?.isValid == false }
     private let pollingLock = WMTLock()
     private var notificationObservers = [NSObjectProtocol]()
-    private let minimumTimePollingInterval = TimeInterval(5)
+    private let minimumTimePollingInterval = 5.0
     
     /// Operation register holds operations in order
     private lazy var operationsRegister = OperationsRegister { [weak self] ops, added, removed in
@@ -405,8 +405,6 @@ class WMTOperationsImpl<T: WMTUserOperation>: WMTOperations, WMTService {
         guard validateActivation(completion) else {
             return
         }
-        
-        let requestStartDate = Date()
         
         networking.post(data: .init(), signedWith: .possession(), to: WMTOperationEndpoints.List<T>.endpoint) { response, error in